# api/docs/swagger.yaml
basePath: /
definitions:
  models.Failed:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.Logout:
    properties:
      message:
        type: string
    type: object
  models.ProfileResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - full_name
    - password
    - username
    type: object
  models.Success:
    properties:
      data: {}
      message:
        type: string
    type: object
  models.Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  models.UpdateProfileRequest:
    properties:
      bio:
        type: string
      email:
        type: string
      full_name:
        type: string
      user_id:
        type: string
    required:
    - email
    - full_name
    type: object
host: localhost:8080
info:
  contact: {}
  description: API service
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login a user with username and password
      parameters:
      - description: Login details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Failed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Failed'
      summary: Login a user
  /logout:
    post:
      description: Logout a user by invalidating token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Logout'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Failed'
      security:
      - BearerAuth: []
      summary: Logout a user
  /profile/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve user profile details
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Failed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Failed'
      summary: Get user profile
    put:
      consumes:
      - application/json
      description: Update user profile details
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Update details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Failed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Failed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Failed'
      summary: Update user profile
  /refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh user token with refresh token
      parameters:
      - description: Refresh token details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Failed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Failed'
      summary: Refresh user token
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with username and password and email
      parameters:
      - description: Registration details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Failed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Failed'
      summary: Register a new user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"